fix(dev): nav dev fiable partout — CSS global, LLM guide en page, masquée des vignettes et captures
Trois problèmes de la barre de nav dev (DevNav) corrigés :

1. Persistance : globals.css était importé page par page ; /dev/renders ne
   l'importait pas → la barre y était rendue SANS ses classes Tailwind
   (invisible, cassée). L'import remonte dans le layout racine (app/layout.tsx),
   les imports par page sont supprimés — plus aucune page ne peut « oublier »
   les styles. Un seul fichier CSS dans le projet → aucun enjeu d'ordre.

2. LLM guide : le lien pointait sur /api/llm-guide (text/markdown brut, sans
   chrome) → le menu disparaissait. Le Markdown est extrait dans
   lib/llm-guide-markdown.ts, partagé entre la route API (inchangée pour les
   agents) et une nouvelle page /[lang]/dev/llm-guide (lecture humaine, sous le
   layout → nav dev conservée, 404 en prod via le middleware existant).

3. Vignettes et captures : les iframes de /dev/renders (Live / Comparer) et la
   story iframe de /dev/components embarquaient chacune leur propre barre.
   DevNav se démonte quand window.self !== window.top. Les scripts de capture
   (renders/generate.mjs, scripts/argos-capture.mjs) masquent [data-devnav]
   pour des captures de référence sans chrome dev.
